Adjusts definition of array model type to use schema_io template

+++ samples/client/petstore/python/src/petstore_api/components/schema/array_of_number_only.py
@@ -25,14 +25,14 @@ class ArrayNumberTuple(
 ArrayNumberTupleInput = typing.Union[
     typing.List[
         typing.Union[
-            float,
-            int
+            int,
+            float
         ],
     ],
     typing.Tuple[
         typing.Union[
-            float,
-            int
+            int,
+            float
         ],
         ...
     ]
